+++ core/src/org/dayaway/duckarena/model/api/ILevel.java
@@ -2,6 +2,6 @@ package org.dayaway.duckarena.model.api;
 
 public interface ILevel {
 
-    long getExp();
+    double getExp();
 
 }
